+++ package/CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### What's Changed
 
+- update default ignore patterns
 - fix plugins being replaced instead of accumulated when composing multiple presets
 - remove redundant double-merge performed by built-in presets